Let's try to be a bit more normal
This first commit adds the usual stuff like require, backbone,
bootstrap, and currently only views the notes.  Note that the database
structure is improved to have a dedicated subject and encrypted
columns, so the content column is no longer mucked with at runtime,
minus decrypting it as needed.  This also means the content can be
sent only when needed, big speed boost.