Fix persistence of encrypted notes
The first problem was that the Encrypted field was not being set or
unset as needed. Additionally the Password field was excluded from
JSON parsing, so ultimately was not being consumed. This resulted in
encrypted notes being persisted in plain text (no password) but still
being flagged as being encrypted. In order to ensure the desired bits
are excluded from being sent back to the frontend a ToJSON method was
added to encapsulate this logic.