Initial -secondary implementation, begin 0.1.0
Note: The UI for this is currently terrible, usually requiring a manual
reload of the page to see what's actually happening (state is hard). So
far this has only been tested on Linux using Syncthing between two
computers.

This is a rough implementation that fixes #64

Testing and a UI is still needed.